Change the type of suio::uio_resid from int to size_t.
git-svn-id: http://llvm.org/svn/llvm-project/safecode/trunk@148674 91177308-0d34-0410-b5e6-96231b3b80d8

+++ runtime/DebugRuntime/PrintfSupport.cpp
@@ -98,7 +98,7 @@ struct suio
   // Number of buffers
   int uio_iovcnt;
   // Total number of bytes to print
-  int uio_resid;
+  size_t uio_resid;
 };
 
 //
@@ -232,9 +232,9 @@ do_output(call_info *c, output_parameter *p, struct suio *uio)
     //
     // Allocate a new string if the old one isn't large enough.
     //
-    if ((size_t) uio->uio_resid > bufsz - pos)
+    if (uio->uio_resid > bufsz - pos)
     {
-      do bufsz *= 2; while ((size_t) uio->uio_resid > bufsz - pos);
+      do bufsz *= 2; while (uio->uio_resid > bufsz - pos);
       dest = (char *) realloc(dest, bufsz);
       if (dest == 0)
         return 1; // Output error
